streaming/requirements.txt
confluent-kafka[avro,schemaregistry]==2.4.0
fastavro==1.9.4
httpx==0.27.0
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
